SLING-10676 - script as used for the initial SECURITY.md addition to all repos

+++ add-security-md.sh
@@ -7,11 +7,20 @@
 # to add or update the SECURITY.md file from the
 # "aggregator" project to all repositories found
 # under it.
+export ME=$(basename $0)
 
 SEC=SECURITY.md
+echo "Adding $SEC to $(basename $PWD)"
+
+if [[ -f $ME ]]
+then
+    echo "Ignoring aggregator folder"
+    exit 0
+fi
+
 git checkout master
 git pull origin master
 cat ../aggregator/$SEC > $SEC
 git add $SEC
 git commit -m "SLING-10676 - add or update $SEC"
-git push origin master
+# TODO git push origin master
